Correction de BUGG: Supprime les () dans le nom des formats video pour que les fichiers apparaissent dans la boite de dialogue de sélection de fichier Changement de comportement: Ajout du codec video h264 "Phone Quality" et du codec video Theora VP3 Changement de comportement: Ajout du format de vidéo OGV (Ogg/Theora) Changement de comportement: Ajout des résolutions HVGA, WVGA, XGA Changement de comportement: Ajout d'un onglet dans la boite de configuration pour la gestion de la base des équipements (à finir) Changement de comportement: Ajout d'une zone d'information dans la barre de status indiquant le n° de la diapo en cours et le nombre de diapo Changement de comportement: Ajout d'un message de confirmation pour les suppressions de slide, de bloque et de plan (désactivable dans la boite de configuration) Changement de comportement: Mémorise l'état des guides magnétique pour les 2 boites slide et recadrage Changement de comportement: Passe le réglage de rotation en valeur à virgule dans la boite de recadrage
git-svn-id: file:///media/werkstatt/064f12d4-5f27-4839-9537-9ffb7939d236/ffDiaporama-Entwicklung-2023/ffdiaporama-svn/trunk@93 f803d417-eb43-4c80-8848-1a7435ce69a7

+++ ffDiaporama/DlgImageCorrection.h
@@ -63,7 +63,7 @@ private slots:
     void            Help();
     void            SetSavedWindowGeometry();
 
-    void            s_RotationEDChanged(int Value);
+    void            s_RotationEDChanged(double Value);
     void            s_XValueEDChanged(double Value);
     void            s_YValueEDChanged(double Value);
     void            s_ZoomValueEDChanged(double Value);
